backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase==2.28.2

# AI / Claude
google-generativeai

# PDF parsing
pymupdf==1.27.2

# HTTP requests
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Async support
anyio==4.3.0

# Date/time handling
python-dateutil==2.8.2
